--- a/tools/theme_dev/__init__.py
+++ b/tools/theme_dev/__init__.py
@@ -1,0 +1,14 @@
+"""Theme development tools package.
+
+This package provides utilities for theme development, testing, and debugging.
+"""
+
+from .theme_watcher import ThemeWatcher
+from .performance_profiler import ThemePerformanceProfiler  
+from .style_inspector import StyleInspector
+
+__all__ = [
+    'ThemeWatcher',
+    'ThemePerformanceProfiler', 
+    'StyleInspector'
+]